+++ packages/express/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @metronome-sh/express
 
+## 7.1.0-next.9
+
+### Patch Changes
+
+- Updated things
+- Updated dependencies
+  - @metronome-sh/runtime@7.1.0-next.9
+  - @metronome-sh/config@7.1.0-next.9
+  - @metronome-sh/node@7.1.0-next.9
+  - @metronome-sh/cli@7.1.0-next.9
+
 ## 7.1.0-next.8
 
 ### Patch Changes