+++ packages/node/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @metronome-sh/node
 
+## 7.1.0-next.9
+
+### Patch Changes
+
+- Updated things
+- Updated dependencies
+  - @metronome-sh/runtime@7.1.0-next.9
+
 ## 7.1.0-next.8
 
 ### Patch Changes